refactor: re-arranged folder structure of transaction edit
