--- src/components/TransactionEdit/CategoryQuickEdit/components/Option/Option.types.ts
@@ -1,5 +1,0 @@
-import type { Category } from '../../../../../types/Category.d';
-
-export interface IProps {
-    category: Category;
-}
