--- src/components/TransactionEdit/SelectOption/SelectOption.types.ts
@@ -1,6 +1,0 @@
-import { ITECTransaction } from '../../../contexts/transactionEditContext';
-
-export interface IProps {
-    idx: number;
-    transaction: ITECTransaction;
-}
