+++ src/components/TransactionEdit/components/CategoryQuickEdit/components/Option/Option.types.ts
@@ -1,0 +1,5 @@
+import type { Category } from '../../../../../../types/Category';
+
+export interface IProps {
+    category: Category;
+}
