+++ src/components/TransactionEdit/components/Table/components/CategorySelector/index.ts
