+++ src/components/TransactionEdit/components/Table/components/Row/Row.types.ts
@@ -1,4 +1,4 @@
-import { ITECTransaction } from '../../../contexts/transactionEditContext';
+import { ITECTransaction } from '../../../../../../contexts/transactionEditContext';
 
 export interface IProps {
     columns: { accessorKey: string; header: string }[];
