+++ src/components/TransactionEdit/components/Table/components/SelectOption/SelectOption.types.ts
@@ -1,0 +1,6 @@
+import { ITECTransaction } from '../../../../../../contexts/transactionEditContext';
+
+export interface IProps {
+    idx: number;
+    transaction: ITECTransaction;
+}
