+++ src/components/TransactionEdit/components/Table/components/TransactionDescription/index.ts
